Bug fix: duplicate UserService
