+++ back/ib/src/main/java/com/ib/service/users/impl/UserService.java
@@ -4,17 +4,17 @@ import com.ib.model.users.User;
 import com.ib.repository.users.IUserRepository;
 import com.ib.service.users.interfaces.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserService implements IUserService {
-
-    private final IUserRepository userRepository;
+public class UserService implements IUserService, UserDetailsService {
 
     @Autowired
-    public UserService(IUserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
+    private IUserRepository userRepository;
+
 
     @Override
     public User findByEmail(String email) {
@@ -24,5 +24,15 @@ public class UserService implements IUserService {
     @Override
     public User get(Integer id) {
         return userRepository.findById(id).orElseGet(null);
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        User user = findByEmail(email);
+        if (user == null) {
+            throw new UsernameNotFoundException(String.format("No user found with email '%s'.", email));
+        } else {
+            return user;
+        }
     }
 }
